Move the fixture folder on the top level
