+++ fixtures/declarationWithoutAssignment.example
