+++ fixtures/failing_jsons/failing_metaphysics.json
