+++ fixtures/global_its.example
