+++ fixtures/metaphysics/partner_show.example
